added printstream controller

+++ src/main/java/net/joshuahughes/smokercontroller/function/Function.java
@@ -1,5 +1,5 @@
 package net.joshuahughes.smokercontroller.function;
 
 public interface Function {
-	public double normalize(double min,double max,double value);
+	public float normalize(float min,float max,float value);
 }
